Revert "메인배너 캐러셀, 카드 컴포넌트 캐러셀 추가"
