+++ src/components/layouts/Carousal.Tsx
@@ -1,0 +1,6 @@
+import React from "react";
+
+const Carousal = () => {
+  return <section>asd</section>;
+};
+export default Carousal;
