+++ src/index.tsx
@@ -1,15 +1,12 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App";
-import { ThemeModeProvider } from "./components/Theme/themeCtx";
 
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
 );
 root.render(
   <React.StrictMode>
-    <ThemeModeProvider>
-      <App />
-    </ThemeModeProvider>
+    <App />
   </React.StrictMode>
 );
